Small fixes
- Added work-around for an incorrect value of
  __STDCPP_DEFAULT_NEW_ALIGNMENT__ on some system/compiler combinations
- Can now add extra linker arguments and can remove compiler and linker
  arguments
- Added documentation for fined-grained SSE/AVX compile options
- "-fstack-clash-protection" is no longer automatically removed when
  compiling with Clang and Python was compiled with GCC
- fixed problem compiling with only SSE version 1
- when getting the project version from Git, light-weight tags are now
  looked at

+++ src/render.cpp
@@ -1589,7 +1589,9 @@ namespace impl {
                 TypeError,
                 "The TriangleBatch instance was pickled with a different batch size. It cannot be loaded here.");
 
-            if(args.size() != 3+item.ctrs->batch_size) THROW_PYERR_STRING(TypeError,"wrong number of arguments");
+            if(static_cast<size_t>(args.size()) != 3+item.ctrs->batch_size) THROW_PYERR_STRING(
+                TypeError,
+                "wrong number of arguments");
 
             auto str = from_pyobject<py::bytes>(args[2]);
             if(static_cast<size_t>(str.size()) != width * (dim+1) * sizeof(float)) {
